refactor code according to comments
