+++ datascope_cache.go
